pyproject.toml
[project]
name = "basespa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.5",
    "dj-rest-auth>=7.1.1",
    "django>=6.0",
    "django-allauth>=65.14.3",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg[binary]>=3.3.3",
    "requests>=2.32.5",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 100
